+++ App.js
@@ -1,56 +1,31 @@
 import { useState, useEffect } from "react";
 import { StatusBar } from "expo-status-bar";
-import { Image, ScrollView, StyleSheet, View } from "react-native";
+import { Image, StyleSheet, View } from "react-native";
 import * as Device from "expo-device";
-
-import bg from "./assets/bg.jpg";
+import bg from "./assets/background.jpg";
 import loading from "./assets/svgs/loading.svg";
-
 import Warnings from "./components/Warnings";
 import Warning from "./components/Warning";
 import CitySelector from "./components/CitySelector";
 import Weather from "./components/Weather";
 import Searcher from "./components/Searcher";
-
 import { getWeather } from "./services/weather";
-
 import {
   checkIfLocationEnabled,
   getCurrentLocation,
   checkNetworkStatus,
   checkAirplaneMode,
 } from "./utils/utils";
-
 import { DEVICE_CODE_DESKTOP } from "./constants";
+import { useWeather } from "./hooks/useWeather";
+import { useApi } from "./hooks/useApi";
+import { useConnectivity } from "./hooks/useConnectivity";
 
 export default function App() {
   const [showCities, setShowCities] = useState(false);
-  const [locationState, setLocationState] = useState({
-    selected: null,
-    enabled: false,
-    permissionGranted: false,
-  });
-  const [apiState, setApiState] = useState({
-    loading: false,
-    error: false,
-    data: null,
-  });
-  const [connectivityState, setConnectivityState] = useState({
-    network: null,
-    airplaneMode: false,
-  });
-
-  const updateLocationState = (updates) => {
-    setLocationState((prev) => ({ ...prev, ...updates }));
-  };
-
-  const updateApiState = (updates) => {
-    setApiState((prev) => ({ ...prev, ...updates }));
-  };
-
-  const updateConnectivityState = (updates) => {
-    setConnectivityState((prev) => ({ ...prev, ...updates }));
-  };
+  const { locationState, updateLocationState } = useWeather();
+  const { apiState, updateApiState } = useApi();
+  const { connectivityState, updateConnectivityState } = useConnectivity();
 
   useEffect(() => {
     checkNetworkStatus().then((status) =>
@@ -66,7 +41,7 @@ export default function App() {
         locationState.selected.longitude
       )
         .then((weather) => {
-          setApiState({
+          updateApiState({
             error: false,
             loading: false,
             data: weather,
@@ -101,7 +76,7 @@ export default function App() {
     <View style={styles.container}>
       <StatusBar style="light" />
       <Image source={bg} blurRadius={10} style={styles.bg} />
-      <ScrollView style={styles.scrollFullContent}>
+      <View style={styles.fullContent}>
         <Searcher
           selectedLocation={locationState.selected}
           setShowCities={setShowCities}
@@ -137,7 +112,7 @@ export default function App() {
             )}
           </View>
         )}
-      </ScrollView>
+      </View>
     </View>
   );
 }
@@ -151,9 +126,9 @@ const styles = StyleSheet.create({
     height: "100%",
     width: "100%",
   },
-  scrollFullContent: {
+  fullContent: {
     width: DEVICE_CODE_DESKTOP === Device.deviceType && "50%",
-    margin: DEVICE_CODE_DESKTOP === Device.deviceType && "auto",
+    margin: "auto",
   },
   loading: { flex: 1, justifyContent: "center", alignItems: "center" },
   imageLoading: { width: 100, height: 100 },
