+++ app.json
@@ -22,7 +22,7 @@
       }
     },
     "web": {
-      "favicon": "./assets/favicon.png"
+      "favicon": "./assets/mosey.jpg"
     },
     "plugins": [
       [
