+++ hooks/useConnectivity.js
@@ -1,0 +1,14 @@
+import { useState } from "react";
+
+export const useConnectivity = () => {
+  const [connectivityState, setConnectivityState] = useState({
+    network: null,
+    airplaneMode: false,
+  });
+
+  const updateConnectivityState = (updates) => {
+    setConnectivityState((prev) => ({ ...prev, ...updates }));
+  };
+
+  return { connectivityState, updateConnectivityState };
+};
